requirements.txt
# 音乐作品管理系统依赖包

# Web框架
Flask==2.3.3

# 数据库连接器
PyMySQL==1.1.0

# 图像处理
Pillow==10.0.1

# HTTP请求库（测试用）
requests==2.31.0

# 其他工具包（通常随Flask安装，但明确列出以确保兼容性）
Werkzeug==2.3.7

# 可选：开发和测试依赖
# pytest==7.4.2
# pytest-flask==1.3.0